+++ frontend/javascripts/oxalis/constants.ts
@@ -338,7 +338,7 @@ const Constants = {
   SCALEBAR_OFFSET: 10,
   OBJECT_ID_STRING_LENGTH: 24,
   REGISTER_SEGMENTS_BB_MAX_VOLUME_VX: 512 * 512 * 512,
-  REGISTER_SEGMENTS_BB_MAX_NO_SEGMENTS: 2000,
+  REGISTER_SEGMENTS_BB_MAX_SEGMENT_COUNT: 2000,
 };
 export default Constants;
